Replace datanucleus.schema.autoCreateAll with autocreate{schema,tables,column}
In HopsHive we have FK constraints pointing to external db (HopsFS)
For unit testing is handy to have derby as backend metadata db testing.
You will still be able to run a more integration testing by changing the
test hive-site.xml configuration to point to real database/fs.
This PR removes the autoCreateAll and replace it with autoCreateSchema,
autoCreateTables, autoCreateColumns to leave out the constraint
creation.